Fixed bugs v1

+++ server/server.py
@@ -5,9 +5,9 @@ import csv
 import io
 
 app = Flask(__name__)
-
 app.secret_key = 'abobaBOOBA' 
 app.config['SESSION_TYPE'] = 'filesystem'
+
 
 DB_CONFIG = {
     'dbname': 'store',
@@ -17,8 +17,10 @@ DB_CONFIG = {
     'port': 5432
 }
 
+
 def get_db_connection():
     return psycopg2.connect(**DB_CONFIG)
+
 
 @app.route('/')
 def login_page():
@@ -29,9 +31,11 @@ def login_page():
         
     return send_from_directory('static', 'login.html')
 
+
 @app.route('/register')
 def register_page():
     return send_from_directory('static', 'register.html')
+
 
 @app.route('/admin')
 def admin_page():
@@ -39,11 +43,13 @@ def admin_page():
         return render_template('admin.html', username=session.get('username'))
     return redirect(url_for('login_page'))
 
+
 @app.route('/user')
 def user_page():
     if 'username' in session and session.get('role_id') == 1:
         return send_from_directory('static', 'userpage.html')
     return redirect(url_for('login_page'))
+
 
 @app.route('/login', methods=['POST'])
 def login():
@@ -79,7 +85,6 @@ def login():
             conn.close()
 
 
-# Обработка регистрации
 @app.route('/register', methods=['POST'])
 def register():
     data = request.json
@@ -99,15 +104,14 @@ def register():
         conn = get_db_connection()
         cur = conn.cursor()
 
-        # Проверяем, существует ли пользователь
         cur.execute("SELECT id FROM users WHERE username = %s", (username,))
         if cur.fetchone():
             return jsonify({'success': False, 'message': 'Пользователь уже существует.'})
 
-        # Хэшируем пароль
+
         hashed_password = generate_password_hash(password)
         role_id = 1
-        # Добавляем пользователя с ролью role_id = 1
+
         cur.execute(
             "INSERT INTO users (username, password_hash, role_id) VALUES (%s, %s, %s)",
             (username, hashed_password, role_id)
@@ -125,16 +129,17 @@ def register():
         if conn:
             conn.close()
 
-# Выход из системы
+
 @app.route('/logout')
 def logout():
     session.clear()
     return redirect(url_for('login_page'))
 
+
 @app.route('/products', methods=['GET'])
 def get_products():
     try:
-        # Получаем параметры фильтрации
+      
         name = request.args.get('name', '').strip()
         city = request.args.get('city', '').strip()
         store = request.args.get('store', '').strip()
@@ -187,16 +192,16 @@ def get_products():
     finally:
         conn.close()
 
+
 @app.route('/store_locations', methods=['GET'])
 def get_store_locations():
     try:
         conn = get_db_connection()
         cur = conn.cursor()
-        # Извлекаем данные из таблицы store_locations
+
         cur.execute("SELECT city, store FROM store_locations")
         rows = cur.fetchall()
 
-        # Преобразуем данные в формат { "город": [номера складов] }
         store_locations = {}
         for city, store_number in rows:
             if city not in store_locations:
@@ -206,6 +211,7 @@ def get_store_locations():
         return jsonify(store_locations)
     finally:
         conn.close()
+
 
 @app.route('/add_product', methods=['POST'])
 def add_product():
@@ -220,7 +226,6 @@ def add_product():
 
     city = data.get('city', '').strip()
 
-    # Валидация данных
     if not name:
         return jsonify({'success': False, 'message': 'Название товара некорректно.'})
     if quantity < 0:
@@ -232,7 +237,6 @@ def add_product():
         conn = get_db_connection()
         cur = conn.cursor()
 
-        # Проверка существования пары (city, store) в таблице store_locations
         cur.execute(
             "SELECT 1 FROM store_locations WHERE city = %s AND store = %s",
             (city, store)
@@ -240,7 +244,6 @@ def add_product():
         if not cur.fetchone():
             return jsonify({'success': False, 'message': 'Указанный город и склад не существуют.'})
 
-        # Вставка товара
         cur.execute(
             "INSERT INTO products (name, quantity, purchase_price, city, store) VALUES (%s, %s, %s, %s, %s)",
             (name, quantity, price, city, store)
@@ -253,7 +256,6 @@ def add_product():
         conn.close()
 
 
-# Удаление товара полностью
 @app.route('/delete_product/<int:product_id>', methods=['POST'])
 def delete_product_completely(product_id):
     try:
@@ -263,12 +265,12 @@ def delete_product_completely(product_id):
         conn.commit()
         return jsonify({'success': True, 'message': 'Товар успешно удалён.'}), 200
     except Exception as e:
-        return jsonify({'success': False, 'message': f'Ошибка сервераS'}), 500
+        return jsonify({'success': False, 'message': 'Невозможно удалить товар'}), 500
     finally:
         if conn:
             conn.close()
 
-# Поиск товара
+
 @app.route('/search_product', methods=['GET'])
 def search_product():
     query = request.args.get('query', '')
@@ -289,6 +291,7 @@ def search_product():
     finally:
         conn.close()
 
+
 @app.route('/modify_product/<int:product_id>', methods=['POST'])
 def modify_product(product_id):
     data = request.json
@@ -302,7 +305,6 @@ def modify_product(product_id):
         conn = get_db_connection()
         cur = conn.cursor()
 
-        # Получаем текущее количество товара
         cur.execute("SELECT quantity FROM products WHERE id = %s", (product_id,))
         result = cur.fetchone()
 
@@ -311,7 +313,6 @@ def modify_product(product_id):
 
         current_quantity = result[0]
 
-        # Рассчитываем новое количество
         if action == 'add':
             new_quantity = current_quantity + quantity
         elif action == 'delete':
@@ -319,44 +320,43 @@ def modify_product(product_id):
             if new_quantity < 0:
                 return jsonify({'success': False, 'message': 'Недостаточное количество товара.'}), 400
 
-        # Обновляем количество в базе данных
         cur.execute("UPDATE products SET quantity = %s WHERE id = %s", (new_quantity, product_id))
         conn.commit()
 
         return jsonify({'success': True, 'message': 'Количество успешно обновлено.'})
     except Exception as e:
-        return jsonify({'success': False, 'message': f'Ошибка сервера: {str(e)}'}), 500
+        return jsonify({'success': False, 'message': 'Ошибка при изменении количества'}), 500
     finally:
         if conn:
             conn.close()
+
 
 @app.route('/update_price/<int:product_id>', methods=['POST'])
 def update_price(product_id):
     data = request.json
     new_price = data.get('price')
 
-    if not new_price or not isinstance(new_price, (int, float)) or new_price <= 0:
+    if not new_price or not isinstance(new_price, (int, float)) or new_price < 0:
         return jsonify({'success': False, 'message': 'Некорректная цена.'}), 400
 
     try:
         conn = get_db_connection()
         cur = conn.cursor()
 
-        # Проверяем, существует ли товар
         cur.execute("SELECT id FROM products WHERE id = %s", (product_id,))
         if not cur.fetchone():
             return jsonify({'success': False, 'message': 'Товар не найден.'}), 404
 
-        # Обновляем цену
         cur.execute("UPDATE products SET purchase_price = %s WHERE id = %s", (new_price, product_id))
         conn.commit()
 
         return jsonify({'success': True, 'message': 'Цена успешно обновлена.'})
     except Exception as e:
-        return jsonify({'success': False, 'message': f'Ошибка сервера: {str(e)}'}), 500
+        return jsonify({'success': False, 'message': 'Ошибка при изменении цены'}), 500
     finally:
         if conn:
             conn.close()
+
 
 @app.route('/change_store/<int:product_id>', methods=['POST'])
 def change_store(product_id):
@@ -371,7 +371,6 @@ def change_store(product_id):
         conn = get_db_connection()
         cur = conn.cursor()
 
-        # Обновляем склад для товара
         cur.execute("""
             UPDATE products
             SET city = %s, store = %s
@@ -381,16 +380,19 @@ def change_store(product_id):
 
         return jsonify({'success': True, 'message': 'Склад успешно обновлен.'})
     except Exception as e:
-        return jsonify({'success': False, 'message': f'Ошибка сервера'}), 500
+        return jsonify({'success': False, 'message': 'Ошибка сервера'}), 500
     finally:
         if conn:
             conn.close()
+
 
 @app.route('/add_role', methods=['POST'])
 def add_role():
     data = request.json
     id = data.get('id')
     name = data.get('name')
+    if (len(name) < 1 or len(name) > 32):
+        return jsonify({'success': False, 'message': 'Название роли должно содержать от 1 до 32 символов.'}), 400
     try:
         conn = get_db_connection()
         cur = conn.cursor()
@@ -399,9 +401,12 @@ def add_role():
             (id, name)
         )
         conn.commit()
-        return '', 201
+        return jsonify({'success': True, 'message': 'Роль успешно добавлена.'}), 201
+    except Exception as a:
+        return jsonify({'success': False, 'message': 'Ошибка при добавлении роли.'}), 500
     finally:
         conn.close()
+
 
 @app.route('/delete_role', methods=['POST'])
 def delete_role():
@@ -409,9 +414,9 @@ def delete_role():
     role_id = data.get('id') 
 
     if not role_id:
-        return jsonify({'error': 'ID роли обязателен'}), 400
-    if (role_id == '777'):
-        return jsonify({'error':'Вы не можете удалить роль admin'}), 400
+        return jsonify({'success': False, 'message': 'ID обязателен.'}), 400
+    if (role_id == 1):
+        return jsonify({'success': False, 'message': 'Вы не можете удалить роль admin.'}), 400
     try:
         conn = get_db_connection()
         cur = conn.cursor()
@@ -419,12 +424,13 @@ def delete_role():
             "DELETE FROM roles WHERE id = %s", (role_id,)
         )
         conn.commit()
-        return '', 204  # Код 204 (No Content), если удаление успешно
+        return jsonify({'success': True, 'message': 'Роль успешно удалена.'}), 201
     except Exception as e:
         conn.rollback()
-        return jsonify({'error': 'Ошибка при удалении роли'}), 500
+        return jsonify({'success': False, 'message': 'Ошибка при удалении роли.'}), 500
     finally:
         conn.close()
+
 
 @app.route('/roles', methods=['GET'])
 def get_roles():
@@ -437,6 +443,7 @@ def get_roles():
     finally:
         conn.close()
 
+
 @app.route('/update_user_role', methods=['POST'])
 def update_user_role():
     data = request.json
@@ -444,29 +451,28 @@ def update_user_role():
     new_role = data.get('role')
 
     if not username or not new_role:
-        return jsonify({"error": "Не указаны все данные"}), 400
+        return jsonify({'success': False, 'message': 'Поля не заполнены.'}), 400
 
     current_user = session.get('username')
     if username == current_user:
-        return jsonify({"error": "Вы не можете изменить свою роль"}), 403
+        return jsonify({'success': False, 'message': 'Нельзя изменять свою роль.'}), 400
 
     try:
         conn = get_db_connection()
         cur = conn.cursor()
 
-        # Проверка существования пользователя
         cur.execute("SELECT * FROM users WHERE username = %s", (username,))
         user = cur.fetchone()
         if not user:
-            return jsonify({"error": "Пользователь не найден"}), 404
+            return jsonify({'success': False, 'message': 'пользователь не найден.'}), 201
 
-        # Обновление роли
         cur.execute("UPDATE users SET role_id = %s WHERE username = %s", (new_role, username))
         conn.commit()
 
-        return '', 200
+        return jsonify({'success': True, 'message': 'Роль успешно обновлена.'}), 201
     finally:
         conn.close()
+
 
 @app.route('/users', methods=['GET'])
 def get_users():
@@ -474,7 +480,6 @@ def get_users():
         conn = get_db_connection()
         cur = conn.cursor()
 
-        # Получение списка пользователей и их ролей
         cur.execute("""
             SELECT u.username, u.role_id, r.name AS role_name FROM users u
             LEFT JOIN roles r ON u.role_id = r.id
@@ -494,26 +499,22 @@ def get_users():
         if conn:
             conn.close()
             
+
 @app.route('/download-csv', methods=['GET'])
 def export_data():
-    format_type = request.args.get('format', 'csv')  # По умолчанию формат CSV
+    format_type = request.args.get('format', 'csv')  
 
     try:
         conn = get_db_connection()
-        if not conn:
-            return jsonify({'error': 'Не удалось установить соединение с базой данных'}), 500
 
         cur = conn.cursor()
-        if not cur:
-            return jsonify({'error': 'Не удалось создать курсор базы данных'}), 500
 
-        # Получение всех данных из таблицы `products`
         try:
             cur.execute("SELECT * FROM products")
             rows = cur.fetchall()
             column_names = [desc[0] for desc in cur.description]
         except Exception as e:
-            return jsonify({'error': f'Ошибка при выполнении запроса к базе данных: {str(e)}'}), 500
+            return jsonify({'error': f'Ошибка при выполнении запроса к базе данных'}), 500
 
         if not rows:
             return jsonify({'error': 'Нет данных для экспорта'}), 404
@@ -523,14 +524,11 @@ def export_data():
                 output = io.StringIO()
                 writer = csv.writer(output)
 
-                # Добавляем заголовок
                 writer.writerow(column_names)
 
-                # Добавляем данные
                 writer.writerows(rows)
                 output.seek(0)
 
-                # Формируем CSV для отправки клиенту
                 return send_file(
                     io.BytesIO(output.getvalue().encode('utf-8')),
                     mimetype='text/csv',
@@ -553,12 +551,19 @@ def export_data():
         except Exception as e:
             print(f'Ошибка при закрытии соединения с базой данных')
 
+
 @app.route('/add_store', methods=['POST'])
 def add_store():
     data = request.json
     city = data.get('city')
-    store = data.get('store')
-
+    try:
+        store = int(data.get('store', -1))
+    except (ValueError, TypeError):
+        return jsonify({'success': False, 'message': 'Номер склада должен быть целым числом.'})
+    
+    if (len(city) < 1 or len(store) > 50):
+        return jsonify({'success': False, 'message': 'Название роли должно содержать от 1 до 50 символов.'}), 400
+    
     try:
        
         conn = get_db_connection()
@@ -568,9 +573,10 @@ def add_store():
             (city, store)
         )
         conn.commit()
-        return '', 201
+        return jsonify({'success': True, 'message': 'Склад успешно добавлен.'}), 201
     finally:
         conn.close()
+
 
 @app.route('/delete_store', methods=['POST'])
 def delete_store():
@@ -579,33 +585,33 @@ def delete_store():
     store = data.get('store')
     
     if not city or not store:
-        return jsonify({'error': 'Поля обязательны для заполнения'}), 400
-    
+        return jsonify({'success': False, 'message': 'Поля не заполнены.'}), 400
+
     try:
         conn = get_db_connection()
         cur = conn.cursor()
-        
-        # Удаление из таблицы products
+
         cur.execute(
-            """
-            UPDATE products
-            SET city = %s
-            WHERE city = %s AND store = %s
-            """,
-            (city+'(удален)', city, store)
+            "SELECT COUNT(*) FROM products WHERE city = %s AND store = %s",
+            (city, store)
         )
-        # Удаление из таблицы store_locations
+        product_count = cur.fetchone()[0]
+
+        if product_count > 0:
+            return jsonify({'success': False, 'message': 'Невозможно удалить склад(он не пуст).'}), 400
+        
         cur.execute(
             "DELETE FROM store_locations WHERE city = %s AND store = %s", (city, store)
         )
-        
+
         conn.commit()
-        return '', 204  # Успешный запрос, но тело ответа пустое
+        return jsonify({'success': True, 'message': 'Склад успешно удалён.'}), 204 
     except Exception as e:
         conn.rollback()
-        return jsonify({'error': 'Ошибка при удалении'}), 500
+        return jsonify({'success': False, 'message': 'Ошибка при удалении склада.'}), 500
     finally:
         conn.close()
+
 
 @app.route('/get_stores', methods=['GET'])
 def get_stores():
@@ -615,7 +621,6 @@ def get_stores():
         cur.execute("SELECT city, store FROM store_locations")
         rows = cur.fetchall()
 
-        # Формируем список складов в формате JSON
         stores = [{'city': row[0], 'store': row[1]} for row in rows]
         return jsonify(stores), 200
     except Exception as e:
@@ -624,18 +629,30 @@ def get_stores():
     finally:
         conn.close()
 
+
 @app.route('/update_product_store', methods=['POST'])
 def update_product_store():
     data = request.json
-    product_id = data['productId']
-    city = data['city']
-    store = data['store']
+    city = data.get('city', '').strip()
+    product_id = data.get('productId')
+    try:
+        store = int(data.get('store', -1))
+        
+    except (ValueError, TypeError):
+        return jsonify({"error": "Неверный номер склада."}), 400
+
+    if not isinstance(product_id, int) or product_id <= 0:
+        return jsonify({"error": "Неверный идентификатор товара."}), 400
+
 
     try:
         conn = get_db_connection()
         cur = conn.cursor()
 
-        # Обновление данных товара
+        cur.execute("SELECT 1 FROM products WHERE id = %s", (product_id,))
+        if cur.fetchone() is None:
+            return jsonify({"error": "Товар с указанным идентификатором не найден."}), 404
+
         cur.execute("""
             UPDATE products
             SET city = %s, store = %s
@@ -645,12 +662,14 @@ def update_product_store():
 
         return jsonify({"message": "Склад обновлен успешно."}), 200
     except Exception as e:
-        return jsonify({"error": "Ошибка"}), 500
+        return jsonify({"error": "Внутренняя ошибка сервера."}), 500
     finally:
         conn.close()
+
 
 @app.route('/health', methods=['GET'])
 def health():
     return "OK", 200
+
 if __name__ == '__main__':
     app.run(debug=True, host="0.0.0.0", port=8080)
